requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
coverage==7.6.2
crispy-bootstrap5==2024.10
Django==5.1.2
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
exceptiongroup==1.2.2
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.29.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.0
PyPDF2==3.0.1
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
rsa==4.9
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
